Split up the acceptance tests to be able to use build tags
This will allow us to run all tests as part of PRs and crons
and also to run a specific set of them for things like smoke testing
providers
